requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
PyJWT==1.7.1
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
PyJWT==1.7.1
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
